share code: go1.8 and go1.7
